+++ plugins/box/system/views/backend/index.view.php
@@ -57,7 +57,15 @@
                     Form::select('site_default_page', $pages_array, Option::get('defaultpage'), array('class' => 'form-control'))
                 );
             ?>
-        </div>
+            </div>
+            <div class="form-group">
+            <?php 
+                echo (
+                    Form::label('site_privacy_policy', __('Privacy policy', 'system')).
+                    Form::select('site_privacy_policy', $pages_array, Option::get('privacy_policy'), array('class' => 'form-control'))
+                );
+            ?>
+            </div>
     </div>
     <div class="col-md-6">
         <h2 class="margin-bottom-1"><?php echo __('System Settings', 'system'); ?></h2>